# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: DevReport AI API
  version: 0.1.0
  description: |
    인증 요청의 IP별 rate limit은 Spring의 `server.forward-headers-strategy=framework`로
    해석된 클라이언트 IP를 사용한다. 배포 시 신뢰된 reverse proxy 또는 ingress만
    `Forwarded`/`X-Forwarded-For`를 정규화해 전달하고, 외부에서 전달된 forwarded header는
    제거해야 한다. Refresh Token은 `refresh_token` HttpOnly 쿠키로만 전달하며,
    Frontend는 인증 요청에 credentials를 포함해야 한다. Backend는 명시된 허용 Origin만
    credentials 요청에 허용하고, 쿠키를 사용하는 인증 요청의 Origin 또는 Referer를 검증한다.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly·SameSite Refresh Token 쿠키. JavaScript에서 읽을 수 없다.
  schemas:
...
    ProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          pattern: '.*\S.*'
          maxLength: 100
...
    ProjectResponse:
      type: object
      required: [id, name, ownerId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FileIdResponse:
      type: object
      required: [fileId]
      properties:
        fileId:
          type: string
          format: uuid
    FileResponse:
      type: object
      required: [id, originalName, contentType, size, createdAt]
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
        contentType:
          type: string
        size:
          type: integer
          format: int64
          minimum: 1
          maximum: 20971520
        createdAt:
          type: string
          format: date-time
    FilePageResponse:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FileResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: [object, "null"]
        timestamp:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /api/projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 프로젝트 상세 조회
      responses:
        "200":
          description: 프로젝트 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트가 없거나 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: 프로젝트 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequest"
      responses:
        "200":
          description: 프로젝트 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: 입력값 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트가 없거나 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: 프로젝트를 휴지통으로 이동
      description: 프로젝트와 파일은 30일간 복구할 수 있으며 이후 자동으로 완전 삭제된다.
      responses:
        "204":
          description: 프로젝트 삭제 성공
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트가 없거나 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/projects/{projectId}/files:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: 프로젝트 파일 업로드
      description: |
        파일당 최대 크기는 20 MiB이다. PDF, DOCX, TXT, MD, ZIP, JPG/JPEG, PNG와 아래
        소스 코드 확장자를 허용하며
        확장자, 요청 MIME, 실제 파일 형식이 모두 일치해야 한다. GIF와 WEBP는 MVP의
        보고서·AI 처리 호환 범위에서 제외되어 FILE_TYPE_NOT_ALLOWED로 거부된다.
        ZIP은 해제 용량 100 MiB와 내부 엔트리 1,000개로 제한하며 ZIP Slip, 환경변수,
        인증서·키, 실행 파일을 차단한다. 소스 코드는 직접 업로드와 ZIP 내부에서 동일하게
        java, kt, py, js, jsx, ts, tsx, html, css, scss,
        sql, xml, json, yaml, yml, md, txt, gradle, properties, toml, go, rs, c, h, cpp, hpp,
        cs, php, rb, swift, dart, vue, svelte 확장자만 별도 디렉터리에 해제한다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: ["file"]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: 파일 업로드 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileIdResponse"
        "400":
          description: 빈 파일, 잘못된 파일 이름, 손상된 PDF·ZIP 또는 ZIP Slip 위험 경로 (FILE_EMPTY, FILE_NAME_INVALID, PDF_INVALID, ZIP_INVALID, ZIP_PATH_INVALID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트가 없거나 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: 파일 크기·PDF 페이지·ZIP 해제·사용자 업로드 quota 초과 (FILE_TOO_LARGE, PDF_PAGE_LIMIT_EXCEEDED, ZIP_LIMIT_EXCEEDED, UPLOAD_STORAGE_QUOTA_EXCEEDED, UPLOAD_FILE_QUOTA_EXCEEDED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: 사용자별 업로드 요청 한도 초과 (RATE_LIMIT_EXCEEDED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: 지원하지 않는 형식 또는 형식 불일치 (FILE_TYPE_NOT_ALLOWED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: 암호화된 PDF 또는 ZIP에 환경변수·인증서·키·실행 파일 포함 (PDF_ENCRYPTED, ZIP_BLOCKED_CONTENT)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 파일 저장소 처리 실패 (FILE_STORAGE_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: 프로젝트 파일 목록 조회
      parameters:
        - name: page
          in: query
          description: 0부터 시작하는 페이지 번호
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: 페이지당 파일 수
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: 생성 시각과 ID 내림차순 파일 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePageResponse"
        "400":
          description: 잘못된 페이지 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트가 없거나 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/projects/{projectId}/files/{fileId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: fileId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 프로젝트 파일 내용 조회
      description: |
        기존 Bearer Token으로 파일을 조회한다. PNG, JPG/JPEG, PDF는 inline으로 응답하고
        나머지 허용 파일은 attachment로 응답한다. Frontend는 인증 요청으로 받은 Blob URL을
        편집기와 출력 페이지에서 사용하며, 파일 URL 자체는 공개하지 않는다.
      responses:
        "200":
          description: 파일 내용 조회 성공
          headers:
            Content-Disposition:
              description: 미리보기 형식은 inline, 그 외 형식은 attachment와 UTF-8 원본 파일명
              schema:
                type: string
            Cache-Control:
              description: 인증 파일의 저장을 막는 no-store 정책
              schema:
                type: string
                example: no-store
            Content-Length:
              description: 저장된 파일 크기
              schema:
                type: integer
                format: int64
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트 또는 파일이 없거나 프로젝트 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 저장 파일이 누락·변조되었거나 파일 저장소 처리 실패 (FILE_STORAGE_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: 프로젝트 파일 삭제
      responses:
        "204":
          description: 파일 메타데이터와 실제 파일 삭제 성공
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트 또는 파일이 없거나 프로젝트 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 파일 저장소 처리 실패 (FILE_STORAGE_ERROR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/report-exports/{exportId}/files/{fileId}:
    get:
      summary: PDF 출력용 이미지 조회
      security: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Render-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 보고서 스냅샷에 포함된 이미지 파일
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: 유효하지 않거나 만료된 출력 토큰·작업·이미지
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...